pyproject.toml
[tool.poetry]
name = "blog-api-redis-graphql"
version = "0.1.0"
description = ""
authors = ["Tobi DEGNON <tobidegnon@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
redis-om = "^0.0.17"
fastapi = "^0.71.0"
uvicorn = {extras = ["standard"], version = "^0.17.0"}
pydantic = "^1.9.0"
python-slugify = "^5.0.2"
strawberry-graphql = {extras = ["fastapi"], version = "^0.93.23"}

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
